change module version constraint
from pessimistic operator to greater than operator.

integrations/github has bumped the major version of the provider
bringing in additional features.

To make this module more accessible and interoperable with other
resources in the new version of the provider, we would have to support
all upward versions until an incompatibility is identified. At which
point, we could consider updating the code to fix the incompatibilty or
add a lesser than operator if the incompatibilty cannot be resolved
within the scope of the code in the repository.